requirements.txt
opencv-python>=4.5.0
pyyaml>=5.4.0
python-rtmidi>=1.4.0
numpy>=1.19.0
